requirements.txt
crewai
langchain-community
python-dotenv
google-generativeai

youtube-search

yfinance

langchain-writer

wolframalpha

wikipedia

wikibase-rest-api-client
mediawikiapi

langchain-vectara

requests

gradio_tools

google-search-results

asknews

langchain-apify